requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
astroid==2.9.3
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
charset-normalizer==2.0.11
colorama==0.4.4
commonmark==0.9.1
discord-colorize==0.0.5
frozenlist==1.3.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
multidict==6.0.2
mypy==0.931
mypy-extensions==0.4.3
platformdirs==2.5.0
Pygments==2.11.2
pylint==2.12.2
PyYAML==6.0
rich==11.2.0
toml==0.10.2
tomli==2.0.1
types-aiofiles==0.8.3
types-PyYAML==6.0.4
typing_extensions==4.0.1
wrapt==1.13.3
yarl==1.7.2
